code refractor increased speed by 25%
